+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Change Log 
 
-# Beta 29 - Grief Defender Port
+
+# Beta 29.1 - Grief Defender Port
 **REQUIRED: Sponge API 7.1 (SF 3682+); Grief Defender 1.12.2-4.3.0.622+; Permissions Plugin (ie. LuckPerms)**
 **OPTIONAL: Nucleus 1.9.0-S7.1+**
+- Fixed a bug where `/is tp` | `/is spawn` would error if an argument was provided
 
-# Beta 29
-**REQUIRED: Sponge API 7.1 (SF 3682+); GP 1.12.2-4.3.0.622+; Permissions Plugin (ie. LuckPerms)**
+# Beta 29 - Grief Defender Port
+**REQUIRED: Sponge API 7.1 (SF 3682+); Grief Defender 1.12.2-4.3.0.622+; Permissions Plugin (ie. LuckPerms)**
 **OPTIONAL: Nucleus 1.9.0-S7.1+**
 - Added support for one player to own multiple islands
   - `/is spawn` & `/is reset` will now display the chest GUI or island list if you have permission to use them on multiple islands